refactor: 메뉴 도메인을 분리하라
