+++ src/main/java/kitchenpos/menu/infra/JpaMenuGroupRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.menu.infra;
 
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
